+++ test_settings.py
@@ -14,6 +14,7 @@ DATABASES = {
 
 INSTALLED_APPS = [
     'django.contrib.sites',
+    'django.contrib.staticfiles',
     'layers'
 ]
 
@@ -35,6 +36,9 @@ STATICFILES_FINDERS = (
 )
 
 STATICFILES_DIRS = (os.path.realpath(os.path.dirname(__file__)) + '/layers/tests/static/',)
+
+STATIC_URL = '/'
+STATIC_ROOT = '/tmp/django-layers/static'
 
 LAYERS = {'layers': ['basic']}